+++ src/App.css
@@ -4,9 +4,7 @@ body {
     padding: 0;
     height: 100%; /* needed for container min-height */
 }
-.header {
-    border: 1px solid red;
-}
+
 .menuLinks {
     background-color: #2559FF;
     font-size: 25px;
@@ -40,7 +38,6 @@ body {
 
 /** search box and autofill **/
 .searchBox {
-    border: 1px solid orange;
     text-align: center;
     margin-left: 20%;
     margin-right: 20%;
@@ -50,7 +47,7 @@ body {
     width: 100%;
     height: 30px;
     font-size: 20px;
-    color: darkblue;
+    padding: 5px;
 }
 .autoFill {
     border: 1px solid green;
@@ -235,4 +232,8 @@ body {
 .uDetails a {
     font-size: 20px;
     color: #0629ff;
+}
+
+.unhandled {
+    margin: 20% 35%;
 }
