+++ src/components/Header.jsx
@@ -1,11 +1,12 @@
 import React from 'react'
+import {Link} from "react-router-dom";
 const Header = () => {
     return (
         <div className='header'>
             <div className='menuLinks'>
-                <div className='homeMenu'><a href='#'>Home</a></div>
-                <a href='#'>About</a>
-                <a href='#'>Contact</a>
+                <div className='homeMenu'><Link to="/">Home</Link></div>
+                <Link to="/about">About</Link>
+                <Link to="/contact">Contact</Link>
             </div>
             <div className='appHeading'>
                 Universities
