+++ src/components/SearchBox.jsx
@@ -3,13 +3,9 @@ const SearchBox = () => {
     return (
         <div className='searchBox'>
             <div className='searchBar'>
-                <input type='text'/>
+                <input type='text' placeholder='Search universities'/>
             </div>
-            <div className='autoFill'>
-                <div>Autofill.. Autofill.. Autofill.. abcd.. </div>
-                <div>Autofill..</div>
-                <div>Autofill..</div>
-            </div>
+
         </div>
 
     )};
